Extract types to separate file.
