+++ packages/eslint-plugin/rules/sort-import-groups/utils.ts
@@ -19,38 +19,19 @@
 /**
  * External dependencies
  */
-import type { AST, Rule } from 'eslint';
 import type * as ESTree from 'estree';
 
-type ESLintSourceCode = ReturnType< Rule.RuleContext[ 'getSourceCode' ] >;
-
 /**
- * Wraps an AST node/comment with the `range` and `loc` fields ESLint always
- * provides at runtime (the ESTree base type marks them optional).
+ * Internal dependencies
  */
-export type Located< T > = T & { range: AST.Range; loc: AST.SourceLocation };
-
-/**
- * `@typescript-eslint/parser` augments `ImportDeclaration` with `importKind`
- * to flag type-only imports. Modeling it inline keeps the rule usable with
- * both the default ESLint parser and the TypeScript parser.
- */
-type TSImportDeclaration = ESTree.ImportDeclaration & {
-	importKind?: 'type' | 'value';
-};
-
-export type ImportNode = Located<
-	TSImportDeclaration | ESTree.VariableDeclaration
->;
-export type AnyNode = Located< ESTree.Node >;
-export type LComment = Located< ESTree.Comment >;
-
-export type DependencyGroup =
-	| 'WordPress dependencies'
-	| 'External dependencies'
-	| 'Internal dependencies';
-
-export type GroupedImports = Record< DependencyGroup, ImportNode[] >;
+import type {
+	AnyNode,
+	DependencyGroup,
+	ESLintSourceCode,
+	GroupedImports,
+	ImportNode,
+	LComment,
+} from './types';
 
 export const WORDPRESS_DEPENDENCIES: DependencyGroup = 'WordPress dependencies';
 export const EXTERNAL_DEPENDENCIES: DependencyGroup = 'External dependencies';
